on darwin, rely on -lglfw, not -lglfw3
unbreaks the `make` process for me on darwin (high sierra 10.13.3).

homebrew seems to treat 'glfw3' as an alias for 'glfw', so installing either of
those package names from homebrew should be ok for the glfw prerequisite.  but
the shared library ends up being called libglfw.3.dylib, so use -lglfw, not
-lglfw3.